+++ src/builtin/module_jit.cpp
@@ -635,7 +635,8 @@ extern "C" {
         if ( tab->isLocked() ) context->throw_error_at(at, "can't insert to a locked table");
         TableHash<KeyType> thh(context,valueTypeSize);
         auto hfn = hash_function(*context, key);
-        return thh.reserve(*tab, key, hfn, at);
+        // TODO Phase 7: widen JIT helper return to int64_t — for now this narrows for tables > INT_MAX.
+        return (int32_t) thh.reserve(*tab, key, hfn, at);
     }
 
     void * das_get_jit_table_at ( int32_t baseType, Context * context, LineInfoArg * at ) {
@@ -658,7 +659,8 @@ extern "C" {
     int32_t jit_table_find ( Table * tab, KeyType key, int32_t valueTypeSize, Context * context ) {
         TableHash<KeyType> thh(context,valueTypeSize);
         auto hfn = hash_function(*context, key);
-        return thh.find(*tab, key, hfn);
+        // TODO Phase 7: widen JIT helper return to int64_t — for now this narrows for tables > INT_MAX.
+        return (int32_t) thh.find(*tab, key, hfn);
     }
 
     void * das_get_jit_table_find ( int32_t baseType, Context * context, LineInfoArg * at ) {
